# backend/internal/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "EpicPanel API",
    "version": "1.0.0",
    "description": "Linux-first hosting control plane API. Session cookie or Bearer session token for users; Bearer epk_ tokens for automation (scoped)."
  },
  "servers": [{"url": "/"}],
  "components": {
    "securitySchemes": {
      "bearer": {"type": "http", "scheme": "bearer"}
    },
...
  },
  "paths": {
...
    "/v1/organizations/{org_id}/websites/{website_id}/domains": {
      "get": {"summary": "List domains (billing+)", "security": [{"bearer": []}], "responses": {"200": {"description": "ok"}}},
      "post": {"summary": "Add alias domain (developer+; globally unique)", "security": [{"bearer": []}], "responses": {"201": {"description": "added"}, "409": {"description": "domain taken"}}}
    },
    "/v1/organizations/{org_id}/domains": {"get": {"summary": "List all domains in the org (SSL dashboard, billing+)", "security": [{"bearer": []}], "responses": {"200": {"description": "ok"}}}},
...
    "/v1/organizations/{org_id}/domains/{domain_id}/ssl": {"post": {"summary": "Set SSL mode {none|selfsigned|letsencrypt} (developer+) -> 202", "security": [{"bearer": []}], "responses": {"202": {"description": "issuing"}}}},
    "/v1/organizations/{org_id}/domains/{domain_id}/verify-dns": {"post": {"summary": "Queue DNS verification (developer+) -> 202", "security": [{"bearer": []}], "responses": {"202": {"description": "queued"}}}},
...
  }
}